services/products_admin/api/apps/products/serializers.py
from rest_framework.serializers import ModelSerializer
from .models import Product

class ProductSerializer(ModelSerializer):
    class Meta:
        model = Product
        fields = "__all__"
        app_label = 'api.products'