services/products_admin/requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
pico==2.0.5
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
